Consolidated af plugins into one
