+++ appfog/lib/appfog-vmc-plugin/help.rb
