+++ appfog/lib/appfog-vmc-plugin/version.rb
@@ -1,3 +1,3 @@
-module AFCLIPull
+module VMCAppfog
   VERSION = "0.0.1".freeze
 end
