--- clone/lib/clone-af-cli-plugin/version.rb
@@ -1,3 +1,0 @@
-module AFCLIClone
-  VERSION = "0.0.0.2".freeze
-end
