--- export/lib/export-af-cli-plugin/version.rb
@@ -1,3 +1,0 @@
-module AFCLIExport
-  VERSION = "0.0.0.2".freeze
-end
